requirements.txt
matplotlib~=3.10.0
numpy~=1.26.4
gmsh~=4.13.1
meshio~=5.3.5
scipy~=1.15.1
h5py~=3.13.0
pyqt6-sip~=13.10.2
pytest~=8.4.2
pandas~=2.2.3